// src/app/models/item-details.model.ts
export class ItemDetailsModel {
    itemType: string = "";
    itemDis: string = "";
    item: string = "";
    dayRate!: number;
    making!: number;
    makingPer!: number;
    purity!: number;
    weight: string ="";
    Price: number = 0;
    soldBack: boolean = false;
    soldBackAmt: number = 0;
}